something about pyright/pydantic changed
This is now the recommended way to set default values. Doing it the
previous way causes pyright to assuming you're accidentally omitting
required values. Doing it this way is somehow able to tell pyright that
these fields are actually optional/have defaults and do not need to be
specified. See [here][0] for details. Change seems to have been
introduced by [this][1].

[0]: https://github.com/samuelcolvin/pydantic/issues/3753#issuecomment-1028693384

[1]: https://github.com/samuelcolvin/pydantic/pull/2721